@@ -9,7 +9,7 @@
     </PropertyGroup>
 
     <PropertyGroup>
-        <TargetFrameworks>net461;net462;net47;net471;net472;net48;netcoreapp3.1;net5.0;net6.0;</TargetFrameworks>
+        <TargetFrameworks>net6.0;net7.0</TargetFrameworks>
         <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
         <SignAssembly>true</SignAssembly>
         <ShouldCreateLogs>True</ShouldCreateLogs>
