@@ -7,7 +7,7 @@
         <SccLocalPath>SAK</SccLocalPath>
     </PropertyGroup>
     <PropertyGroup>
-        <TargetFrameworks>netcoreapp3.1;net5.0</TargetFrameworks>
+        <TargetFrameworks>net7.0;net6.0</TargetFrameworks>
         <AssemblyVersion>1.0.5.230</AssemblyVersion>
         <FileVersion>1.0.5.230</FileVersion>
         <Copyright>Dahl Software Services</Copyright>
