@@ -7,7 +7,7 @@
         <SccLocalPath>SAK</SccLocalPath>
     </PropertyGroup>
     <PropertyGroup>
-        <TargetFrameworks>net461;net462;net47;net471;net472;net48;netcoreapp3.1;net5.0;net6.0;</TargetFrameworks>
+        <TargetFrameworks>net6.0;net7.0;</TargetFrameworks>
         <AssemblyVersion>1.0.6.307</AssemblyVersion>
         <FileVersion>1.0.6.307</FileVersion>
         <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
@@ -18,30 +18,20 @@
         <RepositoryType />
         <PackageTags>ado.net</PackageTags>
         <Description>Database facade for ADO.NET for SQL Server</Description>
-        <PackageReleaseNotes>Updated to work with .NET 4.6x, 4.7x, 4.8, .NET Core 2.1, 3.1, .NET 5.0</PackageReleaseNotes>
+        <PackageReleaseNotes>Updated to work with .NET Core 5.0 and 6.0</PackageReleaseNotes>
         <PackageLicenseExpression></PackageLicenseExpression>
         <PackageLicenseFile>LICENSE.txt</PackageLicenseFile>
     </PropertyGroup>
-    <PropertyGroup Condition="'$(Configuration)|$(TargetFramework)|$(Platform)'=='Debug|net461|AnyCPU'">
+    <PropertyGroup Condition="'$(Configuration)|$(TargetFramework)|$(Platform)'=='Debug|AnyCPU'">
         <CodeAnalysisRuleSet />
         <NoWarn>1701;1702;CA1031;</NoWarn>
     </PropertyGroup>
-    <PropertyGroup Condition="'$(Configuration)|$(TargetFramework)|$(Platform)'=='Release|net461|AnyCPU'">
+    <PropertyGroup Condition="'$(Configuration)|$(TargetFramework)|$(Platform)'=='Release|AnyCPU'">
         <CodeAnalysisRuleSet />
     </PropertyGroup>
 
-    <ItemGroup Condition="'$(TargetFramework)' == 'net5.0'">
-        <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
-		<PackageReference Include="Microsoft.Extensions.Caching.Memory" Version="5.0.0" />
-		<PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="5.0.0" />
-	</ItemGroup>
-	<ItemGroup Condition="'$(TargetFramework)' == 'net6.0'">
-		<PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
-		<PackageReference Include="Microsoft.Extensions.Caching.Memory" Version="5.0.0" />
-		<PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="5.0.0" />
-	</ItemGroup>
-
 	<ItemGroup>
+        <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
         <ProjectReference Include="..\Dahl.Data.Common\Dahl.Data.Common.csproj" />
     </ItemGroup>
     <ItemGroup>
@@ -51,8 +41,6 @@
         </None>
     </ItemGroup>
     <ItemGroup>
-      <PackageReference Include="Microsoft.Extensions.Caching.Memory" Version="6.0.1" />
-      <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="6.0.0" />
-      <PackageReference Include="System.Runtime.Caching" Version="6.0.0" />
+      <PackageReference Include="Microsoft.Extensions.Caching.Memory" Version="7.0.0" />
     </ItemGroup>
 </Project>
