@@ -1,5 +1,4 @@
-﻿#if !NETCOREAPP3_1 && !NET5_0_OR_GREATER
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
@@ -418,4 +417,3 @@ namespace Dahl.Data.Common
         }
     }
 }
-#endif
